Moved Travis related plugin to release package
This concludes moving classes to better packages for now.
