+++ src/main/groovy/org/shipkit/internal/gradle/java/ShipkitJavaPlugin.java
@@ -3,8 +3,8 @@ package org.shipkit.internal.gradle.java;
 import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
-import org.shipkit.internal.gradle.TravisPlugin;
 import org.shipkit.internal.gradle.bintray.BintrayReleasePlugin;
+import org.shipkit.internal.gradle.release.TravisPlugin;
 
 /**
  * Continuous delivery for Java with Travis and Bintray.
