+++ src/main/resources/META-INF/gradle-plugins/org.shipkit.travis.properties
@@ -1,1 +1,1 @@
-implementation-class=org.shipkit.internal.gradle.TravisPlugin
+implementation-class=org.shipkit.internal.gradle.release.TravisPlugin
