+++ src/test/groovy/org/shipkit/internal/gradle/release/TravisPluginTest.groovy
@@ -1,9 +1,8 @@
-package org.shipkit.internal.gradle
+package org.shipkit.internal.gradle.release
 
 import org.gradle.testfixtures.ProjectBuilder
 import org.shipkit.internal.gradle.git.GitBranchPlugin
 import org.shipkit.internal.gradle.git.GitSetupPlugin
-import org.shipkit.internal.gradle.release.ReleaseNeededPlugin
 import spock.lang.Specification
 
 class TravisPluginTest extends Specification {
